Added pin mapping for Pro-Micro, as well as Uno, toggled automatically from build variable __AVR_ATmega32U4__

+++ bootstest/NeoPixelParallel.h
@@ -96,7 +96,7 @@ public:
   void setModeAll();
   void setModeAny();
     
-  void setPinMask(uint8_t mask);
+  void setPinMask(uint16_t mask);
   void setPin(uint8_t p);
     
   void addPixelColor(uint16_t n, uint8_t r, uint8_t g, uint8_t b);
